core: update Invoice services to use BillableCall entity
