+++ library/Ivoz/Kam/Infrastructure/Persistence/Doctrine/TrunksCdrDoctrineRepository.php
@@ -6,7 +6,7 @@ use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Ivoz\Kam\Domain\Model\TrunksCdr\TrunksCdrRepository;
 use Ivoz\Kam\Domain\Model\TrunksCdr\TrunksCdr;
 use Ivoz\Core\Infrastructure\Persistence\Doctrine\Model\Helper\CriteriaHelper;
-use Ivoz\Kam\Infrastructure\Persistence\Doctrine\Traits\GetGeneratorByConditionsTrait;
+use Ivoz\Core\Infrastructure\Persistence\Doctrine\Traits\GetGeneratorByConditionsTrait;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\Validator\Constraints\DateTime;
 
@@ -24,118 +24,6 @@ class TrunksCdrDoctrineRepository extends ServiceEntityRepository implements Tru
     {
         parent::__construct($registry, TrunksCdr::class);
     }
-
-    /**
-     * @param int $invoiceId
-     */
-    public function resetInvoiceId(int $invoiceId)
-    {
-        $qb = $this
-            ->createQueryBuilder('self')
-            ->update($this->_entityName, 'self')
-            ->set('self.invoice', ':nullValue')
-            ->setParameter(':nullValue', null)
-            ->where('self.invoice = :invoiceId')
-            ->setParameter(':invoiceId', $invoiceId);
-
-        return $qb->getQuery()->execute();
-    }
-
-    /**
-     * @param array $ids
-     */
-    public function resetMetered(array $ids)
-    {
-        $qb = $this
-            ->createQueryBuilder('self')
-            ->update($this->_entityName, 'self')
-            ->set('self.metered', ':metered')
-            ->setParameter(':metered', 0)
-            ->where('self.id in (:ids)')
-            ->setParameter(':ids', $ids);
-
-        return $qb->getQuery()->execute();
-    }
-
-    /**
-     * @param array $conditions
-     * @param int $invoiceId
-     * @return mixed
-     */
-    public function setInvoiceId(array $conditions, int $invoiceId)
-    {
-        $qb = $this
-            ->createQueryBuilder('self')
-            ->update($this->_entityName, 'self')
-            ->set('self.invoice', ':invoiceId')
-            ->setParameter(':invoiceId', $invoiceId)
-            ->addCriteria(
-                CriteriaHelper::fromArray($conditions)
-            );
-
-        return $qb->getQuery()->execute();
-    }
-
-    /**
-     * @param int $companyId
-     * @param int $brandId
-     * @param string $startTime
-     * @return mixed
-     */
-    public function countUntarificattedCallsBeforeDate(int $companyId, int $brandId, string $startTime)
-    {
-        $qb = $this->createQueryBuilder('self');
-        $conditions = [
-            ['company', 'eq', $companyId],
-            ['brand', 'eq', $brandId],
-            ['startTime', 'lt', $startTime],
-            ['carrier', 'neq', null],
-            ['carrier', 'neq', ''],
-            ['price', 'isNull'],
-            ['cgrid', 'isNull']
-        ];
-
-        $qb
-            ->select('count(self)')
-            ->addCriteria(
-                CriteriaHelper::fromArray($conditions)
-            );
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
-    /**
-     * @param int $companyId
-     * @param int $brandId
-     * @param string $startTime
-     * @return mixed
-     */
-    public function countUntarificattedCallsInRange(int $companyId, int $brandId, string $startTime, string $endTime)
-    {
-        $qb = $this->createQueryBuilder('self');
-
-        $conditions = [
-            ['company', 'eq', $companyId],
-            ['brand', 'eq', $brandId],
-            ['startTime', 'gt', $startTime],
-            ['endTime', 'lt', $endTime],
-            ['carrier', 'neq', null],
-            ['carrier', 'neq', ''],
-            ['price', 'isNull'],
-            ['cgrid', 'isNull']
-        ];
-
-        $qb
-            ->select('count(self)')
-            ->addCriteria(
-                CriteriaHelper::fromArray($conditions)
-            );
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
-
-
 
     /**
      * This method expects results to be marked as metered as soon as they're used:
